+++ index.html
@@ -62,7 +62,7 @@
       </nav>
     </footer>
     
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
   </html>
